Refactor Mixins into Manager
- AdminActions, UsesQuotaMixin and QuotaParamParser logic moved to QuotaManager
